+++ tests/unit_test/test_resource.py
@@ -29,7 +29,8 @@ class TestResource(unittest.TestCase):
 
     def test_valid_args(self):
         ''' Valid arguments. '''
-        resource = Resource(dim_nodes=PhyDim2(2, 2),
+        resource = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                   origin=PhyDim2(0, 0)),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
@@ -38,15 +39,15 @@ class TestResource(unittest.TestCase):
                                          NodeRegion(dim=PhyDim2(2, 1),
                                                     origin=PhyDim2(0, 1)))
                            )
-        self.assertTupleEqual(resource.dim_nodes, (2, 2), 'dim_nodes')
+        self.assertTupleEqual(resource.node_region.dim, (2, 2), 'node_region')
         self.assertTupleEqual(resource.dim_array, (16, 16), 'dim_array')
         self.assertEqual(resource.size_gbuf, 131072, 'size_gbuf')
         self.assertEqual(resource.size_regf, 512, 'size_regf')
 
-    def test_invalid_dim_nodes(self):
-        ''' Invalid dim_nodes. '''
-        with self.assertRaisesRegexp(TypeError, 'Resource: .*dim_nodes.*'):
-            _ = Resource(dim_nodes=(2, 2),
+    def test_invalid_node_region(self):
+        ''' Invalid node_region. '''
+        with self.assertRaisesRegexp(TypeError, 'Resource: .*node_region.*'):
+            _ = Resource(node_region=PhyDim2(2, 2),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=512,
@@ -59,7 +60,8 @@ class TestResource(unittest.TestCase):
     def test_invalid_dim_array(self):
         ''' Invalid dim_array. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*dim_array.*'):
-            _ = Resource(dim_nodes=PhyDim2(2, 2),
+            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0)),
                          dim_array=(16, 16),
                          size_gbuf=131072,
                          size_regf=512,
@@ -72,7 +74,8 @@ class TestResource(unittest.TestCase):
     def test_invalid_size_gbuf(self):
         ''' Invalid size_gbuf. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*size_gbuf.*'):
-            _ = Resource(dim_nodes=PhyDim2(2, 2),
+            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=(131072,),
                          size_regf=512,
@@ -85,7 +88,8 @@ class TestResource(unittest.TestCase):
     def test_invalid_size_regf(self):
         ''' Invalid size_regf. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*size_regf.*'):
-            _ = Resource(dim_nodes=PhyDim2(2, 2),
+            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
@@ -98,7 +102,8 @@ class TestResource(unittest.TestCase):
     def test_invalid_mem_regions_type(self):
         ''' Invalid mem_regions type. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*mem_regions.*'):
-            _ = Resource(dim_nodes=PhyDim2(2, 2),
+            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
@@ -106,7 +111,8 @@ class TestResource(unittest.TestCase):
                                                  origin=PhyDim2(0, 0))]
                         )
         with self.assertRaisesRegexp(TypeError, 'Resource: .*mem_regions.*'):
-            _ = Resource(dim_nodes=PhyDim2(2, 2),
+            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
@@ -114,7 +120,8 @@ class TestResource(unittest.TestCase):
                                                 origin=PhyDim2(0, 0))
                         )
         with self.assertRaisesRegexp(TypeError, 'Resource: .*mem_regions.*'):
-            _ = Resource(dim_nodes=PhyDim2(2, 2),
+            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
@@ -126,7 +133,8 @@ class TestResource(unittest.TestCase):
     def test_invalid_mem_regions_len(self):
         ''' Invalid mem_regions len. '''
         with self.assertRaisesRegexp(ValueError, 'Resource: .*mem_regions.*'):
-            _ = Resource(dim_nodes=PhyDim2(2, 2),
+            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
@@ -138,13 +146,15 @@ class TestResource(unittest.TestCase):
         ''' Accessor mem_region_src. '''
         nr1 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 0))
         nr2 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 1))
-        resource = Resource(dim_nodes=PhyDim2(4, 4),
+        resource = Resource(node_region=NodeRegion(dim=PhyDim2(4, 4),
+                                                   origin=PhyDim2(0, 0)),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
                             mem_regions=(nr1, nr2))
         self.assertEqual(resource.mem_region_src(), nr1, 'mem_region_src')
-        resource = Resource(dim_nodes=PhyDim2(4, 4),
+        resource = Resource(node_region=NodeRegion(dim=PhyDim2(4, 4),
+                                                   origin=PhyDim2(0, 0)),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
@@ -155,13 +165,15 @@ class TestResource(unittest.TestCase):
         ''' Accessor mem_region_dst. '''
         nr1 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 0))
         nr2 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 1))
-        resource = Resource(dim_nodes=PhyDim2(4, 4),
+        resource = Resource(node_region=NodeRegion(dim=PhyDim2(4, 4),
+                                                   origin=PhyDim2(0, 0)),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
                             mem_regions=(nr1, nr2))
         self.assertEqual(resource.mem_region_dst(), nr2, 'mem_region_dst')
-        resource = Resource(dim_nodes=PhyDim2(4, 4),
+        resource = Resource(node_region=NodeRegion(dim=PhyDim2(4, 4),
+                                                   origin=PhyDim2(0, 0)),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,